razorpay-payment-integration-flask/requirements.txt
asgiref==3.7.2
Flask==2.3.3
Jinja2==3.1.2
python-dotenv==1.0.0
razorpay==1.4.1
